This fixes issue #24 (it makes VarsBase non-copyable).
An object of type VarsBase should not be copied as this will cause problems
with the assignement pointers. I have deleted the copy constructor and
the = operator in VarsBase and have rewritten the core code such that no
copying takes place.

Users should either avoid copying objects of Vars type (recommended) or write
explicit copy constructors and = operators for their objects which redefine the enum
mappings.