Removed divisions by GR_SPACEDIM and replaced by ONE_OVER_GR_SPACEDIM
We often use 1/( (double) GR_SPACEDIM). Not only is this not ideal because of the danger of integer division (hence the
explicit (double) in our code) but it also isn't great from a performance point of view if this is used inside a loop. I
have introduced a #define ONE_OVER_GR_SPACEDIM now and used it in CCZ4.impl.hpp and TensorAlgebra.hpp.

Of course, no old code is broken and 1./GR_SPACEDIM is still absolutely fine.