Adding functionality to store a vector. Details:
Before we could only store 1 value at a time or store an interval of components from a vars object. I have added
functionality to store a 1-tensor of arbitrary size (even if it is not part of a vars object). One gives the start_comp
rather than a Chombo Interval ... I don't fully like this but unfortunately the start and end components of a Chombo Interval are
not compile time constants so there would be not way to check whether the Interval is compatible at compile time.