Only calculating necessary 2nd derivatives:
The new Vars implementation makes it possible to only calculate necessary 2nd derivatives
in a very safe (checked at compile time) way. To this end, a new Diff2Vars struct was
created which holds the variables that require 2nd derivatives.
Some other calls (rhs_equation and calculation of ricci) had to be templated over the d2Vars type as a result.
Only calculating necessary 2nd derivatives leads to roughly ~20% speedup of the RHS.