Added unit tests for the fourth order derivatives.
While the fourth order derivatives are also tested indirectly by
the cpp ChF comparison I thought it would be good to push these unit
tests too because:
a) they are very helpful when writing new derivatives code (e.g.
different order or boundary derivatives for boundary conditions)
b) if there should ever be an error in the derivatives it will be
easier to catch it directly with this unit test.