Add first step condition in SmallDataIO
This change allows the user to set their own condition for the first
step (when the header lines are printed) which means that SmallDataIO
can be called in a different place to specificPostTimeStep. The
WeylExtraction class has been correspondingly updated but old
constructors are kept for backwards compatibility.

The member function definitions have also been moved to a cpp file since
there's no templating.