Modified Ricci tensor calculation
This change should not affect CCZ4 evolution or whenever compute_ricci_Z
is called with the correct value of Z_over_chi but does change
compute_ricci to use the [1st and 2nd] derivatives of h to calculate the
derivative of chris.contracted from its definition rather than using
the derivative of the evolved Gamma (which includes contributions from
the Z vector in the CCZ4 case). As far as I can see this is only used
currently in the calculation of the constraints and the Weyl scalar.

There is a minor change in the BSSN evolution as compute_ricci_Z with
Z_over_chi = 0 will use the calculated chris.contracted rather than the
evolved Gamma in one non-derivative term. I think this is more
conventional anyway.

The tests have been modified accordingly to reflect these changes.