Refactor UserVariables
This is the same as GRChombo/GRChombo#124
* The enum and name definitions of the common CCZ4 variables in
the provided examples have been moved into a common
CCZ4UserVariables.hpp file.
* The type of UserVariables::variable_names has been changed to a more
modern C++ type of std::array<std::string, NUM_VARS>. Note that this is
now const rather than constexpr but I don't think this is an issue.
* A new variable_name_to_enum function which only work with the above
type for variable_names has been defined in the UserVariables namespace
in a new UserVariables.inc.hpp. Since switching to this
will break users' other examples which don't include this file yet, the
old variable_name_to_enum has been temporarily retained in
ChomboParameters. It should be removed in #119.
* A new template function to concatenate std::arrays is defined in a new
ArrayTools namespace (this really ought to be in the standard library)
which is used to concatenate the CCZ4 variable names array with the user
one.