Add new [Matter]Constraints classes
This is similar to the old Constraints classes but has a new
constructor which allows specifying the vars for
* Ham
* Moms
* Ham_abs_terms
* Moms_abs_terms
where the Moms are intervals. If Ham... is < 0, then it is not saved. If
the size of the Moms... intervals is 0, then it is not saved, if it is
1, then the pointwise norm is stored there and if it is 3 then it is
stored as the usual 3-vector.

As an example the ScalarField example has been modified to only have a
single c_Mom var.