Generalize Interpolator classes
This is an attempt to allow the interpolation algorithms (namely Lagrange) to
be used outside of the AMRInterpolator to anyone that needs it and to the
AHFinder. The algorithm we have is prepared for any dimension, but we only use
it for CH_SPACEDIM. If we want to interpolate something else, this should be a
ready to use class. That is the purpose of this commit. The algorithm is
actually "ready" to even do non-uniform grids ('dx' not constant), but changing
that required the structure to change too much.
For these alternative interpolations, the classes 'SimpleArrayBox' and
'SimpleInterpSource' were added, mimicking 'FArrayBox' and 'GRAMRLevel'. They
represent a D-dimensional grid that can be provided based on a flattened 1D
array, potentially with periodic BCs and constant 'dx' (per direction).

Use cases:
1) interpolating the AHFinder AH centers from the 'stats' output when frequency
of solving the AHFinder changes at a restart (1D interpolation)
2) interpolating the AH surface when 'num_points_u' or 'num_points_v' change at
a restart (2D interpolation)
3) interpolating spherically symmetric initial data into the 3D grid (1D
interpolation)